requirements.txt
fastapi== 0.117.1
uvicorn==0.23.2
pandas==2.1.0
python-jose==3.3.0
pydantic==2.3.0
openpyxl==3.1.2
sqlalchemy==2.0.20
pymysql==1.1.0
python-multipart==0.0.6
jdatetime==4.1.1
passlib==1.7.4
python-dotenv==1.0.0